Code analysis fix
